@@ -1,52 +1,29 @@
 {
-  "name": "markets_coinpedia",
+  "name": "markets",
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_OPTIONS='--openssl-legacy-provider' next dev -p 4200",
+    "dev": "next dev -p 4500",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "@analytics/google-analytics": "^0.5.3",
-    "@tinymce/tinymce-react": "^3.12.6",
-    "@visx/axis": "^2.12.2",
-    "@visx/event": "^2.6.0",
-    "@visx/grid": "^2.12.2",
-    "@visx/group": "^2.10.0",
-    "@visx/legend": "^2.10.0",
-    "@visx/mock-data": "^2.1.2",
-    "@visx/responsive": "^2.10.0",
-    "@visx/scale": "^2.2.2",
-    "@visx/shape": "^2.12.2",
-    "@visx/tooltip": "^2.10.0",
-    "analytics": "^0.7.5",
-    "axios": "^0.21.1",
-    "cookie": "^0.4.1",
-    "cross-env": "^7.0.3",
-    "ethers": "^5.3.1",
-    "fs": "0.0.1-security",
-    "highcharts": "^9.3.3",
-    "jquery": "^3.6.0",
-    "js-cookie": "^3.0.1",
+    "@tinymce/tinymce-react": "^4.3.0",
+    "axios": "^1.4.0",
+    "cookie": "^0.5.0",
+    "ethers": "^6.3.0",
+    "js-cookie": "^3.0.5",
     "lightweight-charts": "^4.0.1",
-    "moment": "^2.29.1",
-    "multiselect-react-dropdown": "^2.0.5",
-    "next": "10.2.3",
-    "react": "17.0.2",
-    "react-apexcharts": "^1.4.0",
-    "react-content-loader": "^6.0.3",
-    "react-date-range": "^1.4.0",
-    "react-datetime": "^3.1.1",
-    "react-dom": "17.0.2",
-    "react-dynamic-import": "^1.1.5",
-    "react-image-crop": "^8.6.12",
-    "react-owl-carousel": "^2.3.3",
-    "react-paginate": "^7.1.3",
-    "react-select": "^5.3.2",
-    "react-slick": "^0.29.0",
-    "react-switch": "^6.0.0",
-    "slick-carousel": "^1.8.1",
-    "web3": "^1.3.6"
+    "moment": "^2.29.4",
+    "multiselect-react-dropdown": "^2.0.25",
+    "next": "13.4.1",
+    "react": "18.2.0",
+    "react-content-loader": "^6.2.1",
+    "react-dom": "18.2.0",
+    "react-image-crop": "^10.0.9",
+    "react-paginate": "^8.2.0",
+    "react-select": "^5.7.3",
+    "web3": "^1.9.0"
   }
 }
