@@ -2,27 +2,33 @@
 import React , {useState, useEffect, useRef} from 'react'
 import Axios from 'axios'
 import {API_BASE_URL, graphqlApiKEY, market_coinpedia_url, config, createValidURL, IMAGE_BASE_URL, website_url} from '../components/constants'
-import { useRouter } from 'next/router'
+// import { useRouter } from 'next/navigation'
 import Popupmodal from '../components/popupmodal'
 import Link from 'next/link'
 import JsCookie from "js-cookie" 
+import {TrendingLoader} from '../components/loaders/contentLoader'
+
 
 export default function Details() 
 {   
-    const router = useRouter()
+    // const router = useRouter()
     const [searchBy, setSearchBy] = useState("") 
     const [search_contract_address, set_search_contract_address] = useState("")    
     const [validSearchContract, setvalidContractAddress] = useState("")
     const [err_searchBy, setErrsearchBy] = useState("")
     const [search_container_status, set_search_container_status] = useState(false)
-    const [image_base_url] = useState(IMAGE_BASE_URL + '/tokens/')
-    const [trending_list, set_trending_list] = useState([])
+    const [image_base_url] = useState('https://s2.coinmarketcap.com/static/img/coins/64x64/')
+    
     const [result_showing_type, set_result_showing_type] = useState(1)
     //1:trending & recent searched, 2:Show all tokens, 3:Show all category
-    
-    const [searched_call_status, set_searched_call_status] = useState(false)
+
+    const [default_category_list, set_default_category_list] = useState([])
+    const [default_tokens_list, set_default_tokens_list] = useState([])
     const [searched_tokens_list, set_searched_tokens_list] = useState([])
+
+    const [searched_call_status, set_searched_call_status] = useState(false)
     const [searched_categories_list, set_searched_categories_list] = useState([])
+    const [api_loader,set_api_loader]=useState(false)
 
     var data = []
     if(JsCookie.get("category_search_tokens"))
@@ -31,6 +37,7 @@ export default function Details()
     }
     const [recent_search_list, set_recent_search_list] = useState(data)
     
+    console.log("recent_search_list",recent_search_list)
     const [categories_list, set_categories_list] = useState([])
     const [see_all_trending, set_see_all_trending] = useState(false)
     const [see_all_categories, set_see_all_categories] = useState(false)
@@ -169,13 +176,16 @@ export default function Details()
 
     const trendingCoins=async ()=>
     {
-        const res = await Axios.get(API_BASE_URL+"markets/tokens/trending_assets/", config(""))
+        const res = await Axios.get(API_BASE_URL+"markets/cryptocurrency/trending_assets/", config(""))
         if(res.data)
         {
             if(res.data.status === true)
             { 
+                set_api_loader(true)
                 set_see_all_trending(false)
-                set_trending_list(res.data.message)
+                
+                set_default_tokens_list(res.data.message.tokens_list)
+                set_default_category_list(res.data.message.categories_list)
             } 
         }
     }
@@ -187,7 +197,7 @@ export default function Details()
         {   
             set_searched_call_status(true)
             set_result_showing_type(2)
-            const res = await Axios.get(API_BASE_URL+"markets/tokens/search_assets/"+param, config(""))
+            const res = await Axios.get(API_BASE_URL+"markets/cryptocurrency/search_assets/"+param, config(""))
             if(res.data.status === true)
             {   
                 set_searched_call_status(false)
@@ -230,6 +240,10 @@ export default function Details()
         window.location=market_coinpedia_url+param.token_id
     }
 
+    const Close=()=>{
+        set_search_container_status(false)
+        set_search_contract_address("")
+    }
 return (
     <div>
          {/* <div className="input-group-prepend markets_index">
@@ -243,9 +257,9 @@ return (
                 !search_container_status ?
                 <>
                 <div className="input-group search_filter new_design_serach">
-                    <input onClick={()=>set_search_container_status(true)} type="text" className="form-control search-input-box" placeholder="Search Coin or Address" />
+                    <input onClick={()=>set_search_container_status(true)} type="text" className="form-control search-input-box" placeholder="Search " />
                     <div className="input-group-prepend ">
-                        <span className="input-group-text" onClick={()=>set_search_container_status(true)} ><img src="/assets/img/markets/search.svg" alt="search-box"  width="100%" height="100%"/></span>                 
+                        <span className="input-group-text" onClick={()=>set_search_container_status(true)} ><img src="/assets/img/search_large.svg" alt="search-box"  width="100%" height="100%"/></span>                 
                     </div>
                 </div> 
                 <div className="error">  {err_searchBy}</div>
@@ -263,11 +277,11 @@ return (
                 <div className="search_section" ref={category_search_modal}>
                     <div className="input-group search_input" >
                         <div className="input-group-prepend ">
-                            <span className="input-group-text" ><img className="search-image" src="/assets/img/markets/search.svg" alt="search-box" /></span>                 
+                            <span className="input-group-text" ><img className="search-image" src="/assets/img/search_large.svg" alt="search-box" /></span>                 
                         </div>
                         <input value={search_contract_address} onChange={(e)=> getSearchedResults(e.target.value)} type="text" className="form-control search-input-box" placeholder="Search coin, contract address or exchange" />
                         <div className="input-group-prepend ">
-                            <span className="input-group-text" onClick={()=>set_search_container_status(false)} ><img className="close-image"  src="/assets/img/close.png" alt="search-box" /></span>                 
+                            <span className="input-group-text" onClick={()=>Close()} ><img className="close-image"  src="/assets/img/close.png" alt="search-box" /></span>                 
                         </div>
                     </div>
                     
@@ -276,46 +290,72 @@ return (
                         result_showing_type === 1 ?
                         <>
                             {
-                                (trending_list && trending_list.length > 0) ?
+                                api_loader?
+                                (default_tokens_list && default_tokens_list.length > 0) ?
+                                <>
                                 <div>
-                                    <p>Trending</p>
+                                    <p className='searchbox_titles'>Trending <img className='flame' src="/assets/img/treanding_flame.png" /></p>
                                     <ul className="search-tokens">
                                     {
-                                        trending_list.slice(0,6).map((e,i) =>
+                                        default_tokens_list.slice(0,6).map((e,i) =>
                                             <li onClick={()=>searchByCategoryTokens(e)}>
-                                                {/* <Link> */}
-                                                    <a>
+                                                 <Link href={e.token_id}>
                                                         <div className="media">
                                                             <div className="media-left">
-                                                                <img className="media-object token-img" src={image_base_url+"/"+e.token_image} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
+                                                                <img className="media-object token-img" src={image_base_url+(e.coinmarketcap_id ? e.coinmarketcap_id+".png" : "default.png")} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
                                                             </div>
                                                             <div className="media-body">
-                                                                <p className="media-heading">{e.symbol} <span>{e.token_name}</span></p>
+                                                                <p className="media-heading">{e.token_name} <span>{e.symbol}</span></p>
                                                             </div>
                                                         </div>
-                                                    </a>
-                                                {/* </Link> */}
+                                                </Link>
                                             </li>
                                         )
                                     }
                                     </ul>
-                                </div>    
+                                </div> 
+
+                                <div>
+                                    <p  className='searchbox_titles'>Categories <img className='flame' src="/assets/img/category_list.png" /></p>
+                                    <ul className="search-tokens">
+                                        {
+                                            default_category_list.slice(0,6).map((e,i) =>
+                                                <li>
+                                                    <Link href={"/category/"+e.category_id}>
+                                                        <div className="media">
+                                                            <div className="media-body">
+                                                                <p className="media-heading">{e.category_name}</p>
+                                                            </div>
+                                                            <div className="media-right">
+                                                                <p className='count_token_list'>#{e.total_tokens}</p>
+                                                            </div>
+                                                        </div>
+                                                    </Link>
+                                                </li>
+                                            )
+                                        }
+                                    </ul>
+                                </div>   
+                                </> 
                                 :
                                 ""
+                                :TrendingLoader(1)
                             }
 
                             {
                                 recent_search_list && recent_search_list.length > 0 ?
                                 <>
-                                <p className="mt-4">Recent searches</p>
+                                <p className='searchbox_titles mt-4'>Recent searches</p>
                                 <div>
                                     {
                                         recent_search_list.map((e,i) =>
+                                        <Link href={"/"+e.token_id}>
                                             <div className='recent-searches'>
-                                                <img className="media-object token-img" src={image_base_url+"/"+e.token_image} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
+                                                <img className="media-object token-img" src={image_base_url+"/"+e.token_id} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
                                                 <div className="coin-name">{e.token_name}</div>
                                                 <div className="coin-symbol">{e.symbol}</div>
                                             </div>
+                                             </Link>
                                         )
                                     }
                                     
@@ -332,7 +372,7 @@ return (
                         {
                             (searched_tokens_list && searched_tokens_list.length > 0) ?
                             <div>
-                                <p>Searched Coins</p>
+                                <p className='searchbox_titles'>Searched Coins</p>
                                 <ul className="search-tokens">
                                 {
                                     searched_tokens_list.slice(0,6).map((e,i) =>
@@ -341,10 +381,10 @@ return (
                                                 <a>
                                                     <div className="media">
                                                         <div className="media-left">
-                                                            <img className="media-object token-img" src={image_base_url+"/"+e.token_image} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
+                                                            <img className="media-object token-img" src={image_base_url+(e.coinmarketcap_id ? e.coinmarketcap_id+".png" : "default.png")} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
                                                         </div>
                                                         <div className="media-body">
-                                                            <p className="media-heading">{e.symbol} <span>{e.token_name}</span></p>
+                                                            <p className="media-heading">{e.token_name} <span>{e.symbol}</span></p>
                                                         </div>
                                                     </div>
                                                 </a>
@@ -367,17 +407,23 @@ return (
                         {
                             (searched_categories_list && searched_categories_list.length > 0) ?
                             <div>
-                                <p>Searched Category</p>
+                                <p className='searchbox_titles'>Searched Category</p>
                                 <ul className="search-tokens">
                                 {
                                     searched_categories_list.slice(0,6).map((e,i) =>
-                                    <li><Link href={market_coinpedia_url+"category/"+e.business_id}><a>
-                                        <div className="media">
-                                            <div className="media-body">
-                                                <p className="media-heading">{e.business_name}</p>
-                                            </div>
-                                        </div>
-                                    </a></Link></li>
+                                    
+                                       <li>
+                                            <Link href={"/category/"+e.category_id}>
+                                                <div className="media">
+                                                    <div className="media-body">
+                                                        <p className="media-heading">{e.category_name}</p>
+                                                    </div>
+                                                    <div className="media-right">
+                                                        <p className='count_token_list'>#{e.total_tokens}</p>
+                                                    </div>
+                                                </div>
+                                            </Link>
+                                        </li>
                                     )
                                 }
                                 </ul>
@@ -400,10 +446,10 @@ return (
                             <div>
                                 <div className='row'>
                                     <div className='col-6'>
-                                        <p>Searched Coins</p>
+                                        <p className='searchbox_titles'>Searched Coins</p>
                                     </div>
                                     <div className='col-6 text-right' >
-                                        <a onClick={()=>set_result_showing_type(2)}>Back</a>
+                                        <a onClick={()=>set_result_showing_type(2)}><img className='back_button_search' src="/assets/img/back_icon.png" /></a>
                                     </div>
                                 </div>
                                 <ul className="search-tokens">
@@ -414,7 +460,7 @@ return (
                                                 <a>
                                                     <div className="media">
                                                         <div className="media-left">
-                                                            <img className="media-object token-img" src={image_base_url+"/"+e.token_image} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
+                                                            <img className="media-object token-img" src={image_base_url+(e.coinmarketcap_id ? e.coinmarketcap_id+".png" : "default.png")} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
                                                         </div>
                                                         <div className="media-body">
                                                             <p className="media-heading">{e.symbol} <span>{e.token_name}</span></p>
@@ -439,23 +485,28 @@ return (
                             <div>
                                 <div className='row'>
                                     <div className='col-6'>
-                                        <p>Searched Category</p>
+                                        <p className='searchbox_titles'>Searched Category</p>
                                     </div>
                                     <div className='col-6 text-right' >
-                                        <a onClick={()=>set_result_showing_type(2)}>Back</a>
+                                        <a onClick={()=>set_result_showing_type(2)}><img className='back_button_search' src="/assets/img/back_icon.png" /></a>
                                     </div>
                                 </div>
                                 
                                 <ul className="search-tokens">
                                 {
                                     searched_categories_list.map((e,i) =>
-                                    <li><Link href={market_coinpedia_url+"category/"+e.business_id}><a>
-                                        <div className="media">
-                                            <div className="media-body">
-                                                <p className="media-heading">{e.business_name}</p>
+                                    <li>
+                                        <Link href={"/category/"+e.category_id}>
+                                            <div className="media">
+                                                <div className="media-body">
+                                                    <p className="media-heading">{e.category_name}</p>
+                                                </div>
+                                                <div className="media-right">
+                                                    <p className='count_token_list'>#{e.total_tokens}</p>
+                                                </div>
                                             </div>
-                                        </div>
-                                    </a></Link></li>
+                                        </Link>
+                                    </li>
                                     )
                                 }
                                 </ul>
@@ -470,14 +521,14 @@ return (
                     
                     
                     {/* {
-                        ((trending_list)&&(trending_list.length > 0)&&(!see_all_categories)) ?
+                        ((default_tokens_list)&&(default_tokens_list.length > 0)&&(!see_all_categories)) ?
                         <>
                             <div>
                             <p>Trending</p>
                             <ul>
                                 {
                                     see_all_trending ?
-                                    trending_list.map((e,i) =>
+                                    default_tokens_list.map((e,i) =>
                                     <Link href={website_url+e.token_id}>
                                         <a>
                                             <li>
@@ -496,7 +547,7 @@ return (
                                     :
                                     <>
                                     {
-                                        trending_list.slice(0,6).map((e,i) =>
+                                        default_tokens_list.slice(0,6).map((e,i) =>
                                         <Link href={website_url+e.token_id}>
                                             <a>
                                                 <li>
@@ -527,13 +578,13 @@ return (
                         </>
                         :
                         null
-                    }
+                    } */}
                     
                     {
                         ((categories_list)&&(categories_list.length > 0)&&(!see_all_trending)) ?
                         <>
                             <div className="mt-4">
-                                <p>Category</p>
+                                <p className='searchbox_titles'>Category</p>
                                 <ul>
                                 {
                                     (see_all_categories) ?
@@ -544,7 +595,7 @@ return (
                                                 <img className="media-object token-img" src={image_base_url+"/default.png"} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
                                             </div>
                                             <div className="media-body">
-                                                <p className="media-heading">{e.business_name}</p>
+                                                <p className="media-heading">{e.category_name}</p>
                                             </div>
                                         </div>
                                     </li>
@@ -559,7 +610,7 @@ return (
                                                     <img className="media-object token-img" src={image_base_url+"/default.png"} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={"test"} />
                                                 </div>
                                                 <div className="media-body">
-                                                    <p className="media-heading">{e.business_name}</p>
+                                                    <p className="media-heading">{e.category_name}</p>
                                                 </div>
                                             </div>
                                         </li>
@@ -579,7 +630,7 @@ return (
                         </>
                         :
                         null
-                    } */}
+                    }
                             
 
                         
