@@ -1,5 +1,87 @@
+import moment from 'moment'
 
 const graphQlURL = "https://graphql.bitquery.io/"
+
+const cmc_graph_ranges = [
+  {
+    time_name : "1D",
+    intervals: "10m",
+    count:"144"
+  },
+  {
+    time_name : "1W",
+    intervals: "1h",
+    count:"168"
+  },
+  {
+    time_name : "1M",
+    intervals: "4h",
+    count:"180"
+  },
+  {
+    time_name : "3M",
+    intervals: "1d",
+    count:"90"
+  },
+  {
+    time_name : "6M",
+    intervals: "1d",
+    count:"180"
+  },
+  {
+    time_name : "1Y",
+    intervals: "2d",
+    count:"183"
+  },
+  {
+    time_name : "All",
+    intervals: "3d",
+    count:"243"
+  }  
+]
+
+const cmcArrageGraphData = async (data)=>
+{
+  var final_array  = []
+  var volume_array  = []
+  var market_cap_array  = []
+  
+  var low_value = data[0].quote.USD.price
+  var high_value = 0
+  for(let run of data)
+  { 
+      var create_obj = {}
+      create_obj['time'] = await  Math.floor((new Date(run.timestamp)).getTime()/1000)
+      create_obj['value'] = run.quote.USD.price
+
+      var create_obj2 = {}
+      create_obj2['time'] = await  Math.floor((new Date(run.timestamp)).getTime()/1000)
+      create_obj2['value'] = run.quote.USD.volume_24h
+
+      var create_obj3 = {}
+      create_obj3['time'] = await  Math.floor((new Date(run.timestamp)).getTime()/1000)
+      create_obj3['value'] = run.quote.USD.market_cap
+
+      if(run.quote.USD.price > high_value)
+      {
+        high_value = run.quote.USD.price
+      }
+
+      if(run.quote.USD.price < low_value)
+      {
+        low_value = run.quote.USD.price
+      }
+
+      await final_array.push(create_obj)
+      await volume_array.push(create_obj2)
+      await market_cap_array.push(create_obj3)
+  }
+  const base_price = await ((low_value+high_value)/2).toFixed(0)
+
+  return {final_array, volume_array,market_cap_array, base_price, low_value, high_value}
+}
+
+
 const fromNToDate = (datetime,customstartdate,customenddate)=>
 {
     var from_date = new Date();
@@ -74,4 +156,4 @@ const fromNToDate = (datetime,customstartdate,customenddate)=>
     return {fromDate:fromDate, toDate:toDate}
 }
 
-module.exports = { graphQlURL, fromNToDate }
+module.exports = { graphQlURL, fromNToDate, cmcArrageGraphData, cmc_graph_ranges}
