@@ -1,8 +1,9 @@
-module.exports = {
-    webpack5: false,
+const nextConfig = {
     trailingSlash: true,
     i18n: {
       locales: ["en"],
       defaultLocale: "en",
-    },
+    }
 }
+
+module.exports = nextConfig
