@@ -1,88 +1,101 @@
 import { createChart, ColorType } from 'lightweight-charts';
 import React, { useEffect, useRef } from 'react';
 import Chart from '../components/lightWeightChart'
+import { graphQlURL, fromNToDate, cmcArrageGraphData,cmc_graph_ranges} from '../components/tokenDetailsFunctions' 
 
 
 
-export default function Demo() {
+export default function Demo({data}) {
 
     const chartContainerRef = useRef()
 
+    console.log("cmc_graph_ranges",cmc_graph_ranges)
 
-    useEffect(() => {
+    useEffect( () => {
+        plotGraph()
+       
 
-        const initialData = [
-            { time:  1556877600 , value: 19103293.0 },
-            { time:  1556877700 , value: 21737523.0 },
-            { time:  1556877800 , value: 29328713.0 },
-            { time:  1556877900 , value: 37435638.0 },
-            { time:  1556878000 , value: 25269995.0 },
-            { time:  1556878100 , value: 24973311.0 },
-            { time:  1556878200 , value: 22103692.0 },
-            { time:  1556878300 , value: 25231199.0 },
-            { time:  1556878400 , value: 24214427.0 },
-            { time:  1683714840 , value: 22533201.0 },
-            // { time: "2018-11-02", value: 14734412.0 },
-            // { time: "2018-11-05", value: 12733842.0 },
-            // { time: "2018-11-06", value: 12371207.0 },
-            // { time: "2018-11-07", value: 14891287.0 },
-            // { time: "2018-11-08", value: 12482392.0 },
-            // { time: "2018-11-09", value: 17365762.0 },
-            // { time: "2018-11-12", value: 13236769.0 },
-            // { time: "2018-11-13", value: 13047907.0 },
-            // { time: "2018-11-14", value: 18288710.0 },
-            // { time: "2018-11-15", value: 17147123.0 },
-            // { time: "2018-11-16", value: 19470986.0 },
-            // { time: "2018-11-19", value: 18405731.0 },
-            // { time: "2018-11-20", value: 22028957.0 },
-            // { time: "2018-11-21", value: 18482233.0 },
-            // { time: "2018-11-23", value: 7009050.0 },
-            // { time: "2018-11-26", value: 12308876.0 },
-            // { time: "2018-11-27", value: 14118867.0 },
-            // { time: "2018-11-28", value: 18662989.0 },
-            // { time: "2018-11-29", value: 14763658.0 },
-            // { time: "2018-11-30", value: 31142818.0 },
-            // { time: "2018-12-03", value: 27795428.0 },
-            // { time: "2018-12-04", value: 21727411.0 },
-            // { time: "2018-12-06", value: 26880429.0 },
-            // { time: "2018-12-07", value: 16948126.0 },
-            // { time: "2018-12-10", value: 16603356.0 },
-            // { time: "2018-12-11", value: 14991438.0 },
-            // { time: "2018-12-12", value: 18892182.0 },
-            // { time: "2018-12-13", value: 15454706.0 },
-            // { time: "2018-12-14", value: 13960870.0 },
-            // { time: "2018-12-17", value: 18902523.0 },
-            // { time: "2018-12-18", value: 18895777.0 },
-            // { time: "2018-12-19", value: 20968473.0 },
-            // { time: "2018-12-20", value: 26897008.0 },
-            // { time: "2018-12-21", value: 55413082.0 },
-            // { time: "2018-12-24", value: 15077207.0 },
-            // { time: "2018-12-26", value: 17970539.0 },
-            // { time: "2018-12-27", value: 17530977.0 },
-            // { time: "2018-12-28", value: 14771641.0 },
-            // { time: "2018-12-31", value: 15331758.0 },
-            // { time: "2019-01-02", value: 13969691.0 },
-            // { time: "2019-01-03", value: 19245411.0 },
-            // { time: "2019-01-04", value: 17035848.0 },
-            // { time: "2019-01-07", value: 16348982.0 },
-            // { time: "2019-01-08", value: 21425008.0 },
-            // { time: "2019-01-09", value: 18136000.0 },
-            // { time: "2019-01-10", value: 14259910.0 },
-            // { time: "2019-01-11", value: 15801548.0 },
-            // { time: "2019-01-14", value: 11342293.0 },
-            // { time: "2019-01-15", value: 10074386.0 },
-            // { time: "2019-01-16", value: 13411691.0 },
-            // { time: "2019-01-17", value: 15223854.0 },
-            // { time: "2019-01-18", value: 16802516.0 },
-            // { time: "2019-01-22", value: 18284771.0 },
-        ];
+    }, [])
 
-        console.log("initialData",initialData)
+    const plotGraph = async () =>
+    {
+        
+
+        const {final_array,volume_array, base_price, low_value, high_value} = await cmcArrageGraphData(data.data.ETH[0].quotes)
+
+
+        console.log("low_value", low_value)
+        console.log("final_array", final_array)
+        console.log("high_value", high_value)
+        // const initialData = [
+        //     {
+        //         "time": 1683799440,
+        //         "value": 1820.9087409897747,
+        //         "_internal_originalTime": 1683799440
+        //     },
+        //     {
+        //         "time": 1683799740,
+        //         "value": 1820.8389634376051,
+        //         "_internal_originalTime": 1683799740
+        //     },
+        //     {
+        //         "time": 1683800040,
+        //         "value": 1821.372533803982,
+        //         "_internal_originalTime": 1683800040
+        //     },
+        //     {
+        //         "time": 1683800340,
+        //         "value": 1821.4971115131334,
+        //         "_internal_originalTime": 1683800340
+        //     },
+        //     {
+        //         "time": 1683800640,
+        //         "value": 1822.306387552131,
+        //         "_internal_originalTime": 1683800640
+        //     },
+        //     {
+        //         "time": 1683800940,
+        //         "value": 1822.5122819899095,
+        //         "_internal_originalTime": 1683800940
+        //     },
+        //     {
+        //         "time": 1683801240,
+        //         "value": 1821.7232531232303,
+        //         "_internal_originalTime": 1683801240
+        //     },
+        //     {
+        //         "time": 1683801540,
+        //         "value": 1823.7775303172766,
+        //         "_internal_originalTime": 1683801540
+        //     },
+        //     {
+        //         "time": 1683801840,
+        //         "value": 1824.9184608645571,
+        //         "_internal_originalTime": 1683801840
+        //     },
+        //     {
+        //         "time": 1683802140,
+        //         "value": 1826.2296227176255,
+        //         "_internal_originalTime": 1683802140
+        //     },
+        //     {
+        //         "time": 1683802440,
+        //         "value": 1825.471658770417,
+        //         "_internal_originalTime": 1683802440
+        //     },
+        //     {
+        //         "time": 1683802740,
+        //         "value": 1824.4202477834692,
+        //         "_internal_originalTime": 1683802740
+        //     }
+        // ]
+
+        // console.log("initialData",initialData)
         const chart = createChart(chartContainerRef.current, {
             layout: {
                 background: { type: ColorType.Solid, color: "white" },
             },
-            width: 1000,
+            width: chartContainerRef.current.clientWidth,
             height: 400,
         })
 
@@ -109,16 +122,18 @@ export default function Demo() {
         })
 
         const baselineSeries = chart.addBaselineSeries({
-            baseValue: { type: 'price', price: 22533201 },
+            baseValue: { type: 'price', price: base_price },
+            priceLineVisible: false,
             topLineColor: 'rgba( 38, 166, 154, 1)', topFillColor1: 'rgba( 38, 166, 154, 0.68)',
             topFillColor2: 'rgba( 38, 166, 154, 0.05)', bottomLineColor: 'rgba( 239, 83, 80, 1)',
             bottomFillColor1: 'rgba( 239, 83, 80, 0.05)', bottomFillColor2: 'rgba( 239, 83, 80, 0.68)'
         });
-        baselineSeries.setData(initialData)
+        baselineSeries.setData(final_array)
         chart.timeScale().fitContent();
 
         chart.priceScale("left").applyOptions({
             borderColor: '#ccc4c4 ',
+            autoScale: false,
 
         });
 
@@ -126,10 +141,48 @@ export default function Demo() {
             borderColor: '#ccc4c4',
         });
 
-        window.addEventListener('resize', () => {
-            chart.resize(window.innerWidth, window.innerHeight);
+       
+        const avgPriceLine = {
+            price: parseFloat(base_price),
+            color: '#cfd6e3',
+            // lineWidth: 2,
+            // lineStyle: 1, // LineStyle.Dotted
+            axisLabelVisible: true,
+            
+            // title: 'ave price',
+        };
+
+        baselineSeries.createPriceLine(avgPriceLine);
+//      priceLine.applyOptions({
+//     price: 90.0,
+//     color: 'red',
+//     lineWidth: 3,
+//     lineStyle: LightweightCharts.LineStyle.Dashed,
+//     axisLabelVisible: false,
+//     title: 'P/L 600',
+// });
+
+        const volumeSeries = chart.addHistogramSeries({
+            color: '#cfd6e3',
+            
+            priceFormat: {
+                type: 'volume',
+            },
+            priceScaleId: '', // set as an overlay by setting a blank priceScaleId
+            // set the positioning of the volume series
+            scaleMargins: {
+                top: 0.9, // highest point of the series will be 70% away from the top
+                bottom: 0,
+            },
+        });
+        volumeSeries.priceScale().applyOptions({
+            scaleMargins: {
+                top: 0.9, // highest point of the series will be 70% away from the top
+                bottom: 0,
+            },
         });
 
+        volumeSeries.setData(volume_array)
 
         console.log("chartContainerRef", chart)
         const toolTipWidth = 80;
@@ -158,12 +211,22 @@ export default function Demo() {
             } else {
                 // time will be in the same format that we supplied to setData.
                 // thus it will be YYYY-MM-DD
+
                 const dateStr = param.time;
                 toolTip.style.display = 'block';
                 const data = param.seriesData.get(baselineSeries);
+                const data2 = param.seriesData.get(volumeSeries);
+                console.log("data",data2)
+                const volume = data2.value !== undefined ? data2.value : data2.close;
+                console.log("data",volume)
+                
                 const price = data.value !== undefined ? data.value : data.close;
+                // const volume = data!== undefined ? data.value : data.close;
                 toolTip.innerHTML = `<div style="color: ${'rgba( 38, 166, 154, 1)'}">ABC Inc.</div><div style="font-size: 24px; margin: 4px 0px; color: ${'black'}">
             ${Math.round(100 * price) / 100}
+            </div>
+            <div style="font-size: 24px; margin: 4px 0px; color: ${'black'}">
+            ${volume}
             </div><div style="color: ${'black'}">
             ${dateStr}
             </div>`;
@@ -182,10 +245,22 @@ export default function Demo() {
                 toolTip.style.top = top + 'px';
             }
         });
-        return () => [chart.remove()]
 
+        const handleResize=()=>{
+            chart.applyOptions({
+                width:chartContainerRef.current.clientWidth,
+                // height:400
+            })
+        }
 
-    }, [])
+        window.addEventListener('resize', handleResize)
+        
+        return () => {chart.remove();
+            window.removeEventListener('resize', handleResize)
+            }
+
+    }
+
     return (
         <>
             <div ref={chartContainerRef}>
@@ -196,4 +271,28 @@ export default function Demo() {
 
     );
 
+}
+
+
+export async function getServerSideProps({query,req}) 
+{
+    const req_config = {
+        "mode": 'no-cors',
+        "method": 'get',
+        "headers": {
+            "Content-Type": "application/json"
+        }
+    }
+    // cmc_graph_ranges
+   const tokenQuery = await fetch("https://pro-api.coinmarketcap.com/v3/cryptocurrency/quotes/historical?CMC_PRO_API_KEY=e671363b-4b20-42d7-80fe-32fb2ee8a88b&symbol=eth&count=96&interval=15m", req_config)
+   if(tokenQuery) 
+   {    
+     const tokenQueryRun = await tokenQuery.json()
+     return { props: { data: tokenQueryRun, errorCode: false} }
+   }
+   else 
+   {
+    return { props: { errorCode: true, data:{} } }
+   }
+  // return { props: {userAgent:userAgent, config:config(user_token), user_token:user_token,category_id:category_id}}
 }
