@@ -1,9 +1,9 @@
 
-import React, {useEffect, useState} from 'react'
+import React, {useEffect, useState, useRef} from 'react'
 import Link from 'next/link' 
 import cookie from "cookie"
 import ReactPaginate from 'react-paginate'  
-import { API_BASE_URL, roundNumericValue, config, separator, app_coinpedia_url, IMAGE_BASE_URL, market_coinpedia_url, graphqlApiKEY, count_live_price, Logout} from '../components/constants' 
+import { API_BASE_URL, roundNumericValue, config, separator, app_coinpedia_url, IMAGE_BASE_URL, market_coinpedia_url, strLenTrim, count_live_price, Logout} from '../components/constants' 
 import Axios from 'axios'  
 import Head from 'next/head'
 import SearchContractAddress from '../components/searchContractAddress'
@@ -12,61 +12,68 @@ import TableContentLoader from '../components/loaders/tableLoader'
 import moment from 'moment'
 import WatchList from '../components/watchlist'
 import Select from 'react-select'
-import { useRouter } from 'next/router'
 
-'use client';
+// import Select from 'react-select'
+import { useRouter } from 'next/navigation'
+
 export default function Companies({user_token, config, userAgent})
 { 
     const router = useRouter()
-    const { active_category_tab } = router.query
+    const myRef = useRef(null)
+    //const { active_category_tab } = router.query
+    const active_category_tab = ''
     const [tokens_list, set_tokens_list] = useState([]) 
-    const [voting_ids, setvoting_ids] = useState([])  // commented
-    const [watchlist, set_watchlist] = useState([])
     const [currentPage, setCurrentPage] = useState(0)
     const [per_page_count, set_per_page_count] = useState(100)
     const [pageCount, setPageCount] = useState(0)
     const [sl_no, set_sl_no]=useState(0)
     const [firstcount, setfirstcount] = useState(1)
     const [finalcount, setfinalcount] = useState(per_page_count)
-    const [selectedPage, setSelectedPage] = useState(0) 
     const [image_base_url] = useState('https://s2.coinmarketcap.com/static/img/coins/64x64/')
     const [count, setCount]=useState(0)
-    const [voting_status, set_voting_status] = useState(false)
     const [loader_status, set_loader_status] = useState(false)
-    const [handleModalVote, setHandleModalVote] = useState(false)
-    const [total_votes, set_total_votes] = useState()
-    const [token_id, set_Token_id] = useState("")
-    const [vote_id, set_vote_id] = useState("")
-    const [item, set_item] = useState("")
-    const [voting_message, set_voting_message] = useState("")
     const [all_tab_status, set_all_tab_status] = useState((active_category_tab > 0) ? 0 : 1)
     const [watchlist_tab_status, set_watchlist_tab_status] = useState("")
-    const [search_title, set_search_title] = useState("")  
-
+    const [search_title, set_search_title] = useState("")
     const [category_list, set_category_list] = useState([]) 
-    const [other_category_list, set_other_category_list] = useState([])
+    const [category_id, set_category_id] = useState("") 
+    const [other_category_list, set_other_category_list] = useState([]) 
     const [category_name, set_category_name] = useState("")
-    const [category_row_id, set_category_row_id] = useState((active_category_tab > 0) ? active_category_tab : "") 
+    const [search_category_value, set_search_category_value] = useState("")
+    const [category_status, set_category_status] = useState(false)
+    const [searchParam]=useState(["category_name"])
+    const category_ref = useRef()
     
-    const [business_tab_status, set_business_tab_status] = useState((active_category_tab > 0) ? 2 :"")
     
-    useEffect(async ()=>
+
+    useEffect(()=>
     {  
         tokensList({selected : 0})
-        voteIds()
-        watchListIds()
-        mainUniqueCategory()
-    },[per_page_count, search_title, category_row_id]) 
+        getCategoryList()
+
+        let handler = (e) => {
+          if (!category_ref?.current?.contains(e.target)) {
+            set_category_status(false)
+          }
+        }
+
+        document.addEventListener("mousedown", handler)
+
+          return () => {
+              document.removeEventListener("mousedown", handler)
+          }
+    },[per_page_count, search_title, category_id]) 
 
 
-    const mainUniqueCategory = () =>
+    const getCategoryList = () =>
     {  
-        Axios.get(API_BASE_URL+"markets/tokens/unique_categories", config).then(res=>
+        Axios.get(API_BASE_URL+"markets/cryptocurrency/unique_categories", config).then(res=>
         { 
             if(res.data)
             {      
                 set_category_list(res.data.message)
-                otherUniqueCategory(res.data.other_array)
+                set_other_category_list(res.data.message)
+                // otherUniqueCategory(res.data.other_array)
             } 
         })
     }
@@ -75,13 +82,38 @@ export default function Companies({user_token, config, userAgent})
     {  
         var listData = res
         var list = []
-        for(const i of listData)
+        for(let i of listData)
         {
             list.push({value: i.category_id, label: i.category_name})  
         }
         set_other_category_list(list)
     }
 
+const searchTags = (param_value) => 
+{
+   set_search_category_value(param_value)
+    if (param_value) {
+      var searched_list = other_category_list.filter(
+        // createFilter(param_value, ["category_name"])
+        (item) => {
+          return searchParam.some((newItem) => {
+               return (
+                   item[newItem]
+                       .toString()
+                       .toLowerCase()
+                       .indexOf(param_value) > -1
+                       
+               )
+           })
+       }
+      );
+     set_category_list(searched_list);
+     set_category_status(true);
+    } else {
+      set_category_list(other_category_list);
+    }
+    // console.log(param_value);
+}
     const tokensList = async (page) =>
     {  
         let current_pages = 0 
@@ -89,8 +121,9 @@ export default function Companies({user_token, config, userAgent})
         {
             current_pages = ((page.selected) * per_page_count) 
         } 
-
-        const res = await Axios.get(API_BASE_URL+"markets/cryptocurrency/list/"+current_pages+'/'+per_page_count+"?search="+search_title+"&category_row_id="+category_row_id, config)
+        // myRef.current.scrollIntoView()
+        set_loader_status(false)  
+        const res = await Axios.get(API_BASE_URL+"markets/cryptocurrency/list/"+current_pages+'/'+per_page_count+"?search="+search_title+"&category_id="+category_id, config)
         if(res.data)
         {
             if(res.data.status === true)
@@ -103,6 +136,7 @@ export default function Companies({user_token, config, userAgent})
                 setCurrentPage(page.selected)
                 setfirstcount(current_pages+1)
                 setCount(res.data.count)
+                // set_watchlist_status(res.da)
                 //setfinalcount(parseInt(current_pages)+parseInt(per_page_count))
                 const presentPage = page.selected+1
                 const totalcompany = res.data.count
@@ -113,117 +147,53 @@ export default function Companies({user_token, config, userAgent})
                 }
                 const final_count=sadf
                 setfinalcount(final_count)
+                
             } 
         }
     }
 
-    const voteIds = () =>
+    const addToWatchlist = async (param_token_id) =>
     {
-        Axios.get(API_BASE_URL+"markets/tokens/voting_ids", config).then(res=>
-        { 
-            if(res.data.status)
-            {
-                setvoting_ids(res.data.voting_ids)
-            }
-        })
-    }
-
-    const watchListIds = () =>
-    {
-        Axios.get(API_BASE_URL+"markets/tokens/watchlist_ids", config).then(res=>
-        { 
-            if(res.data.status)
-            {
-            set_watchlist(res.data.message)
-            }
-        })
-    }
-    
-    const ModalVote=(token_id,status,_id,item)=> 
-    {  
-        setHandleModalVote(!handleModalVote) 
-        set_voting_status(status)
-        set_Token_id(token_id)
-        set_vote_id(_id)
-        set_item(item)
-    }
-  
-    const vote = (param) =>
-    {
-      
-      if(param == 1)
-      {
-        Axios.get(API_BASE_URL+"markets/listing_tokens/save_voting_details/"+token_id, config)
-        .then(res=>
-        { 
-          if(res.data.status === true) 
-          {
-            
-            var testList = tokens_list
-            var result = testList.filter(obj => {
-              return obj._id === vote_id
-            })
-            var testObj = result ? result[0] : "" 
-            var test_total_votes = testObj.total_votes+1
-            testObj['total_votes'] = test_total_votes
-            testList[item] = testObj
-            set_tokens_list(testList)
-            voting_ids.push(vote_id)
-            set_voting_message(res.data.message) 
-            setHandleModalVote(!handleModalVote)
-          }
-        })
-      }
-      else
-      {
-        Axios.get(API_BASE_URL+"markets/listing_tokens/remove_voting_details/"+token_id, config)
-        .then(res=>
-        { 
-          if(res.data.status === true) 
-          {
-            var testList = tokens_list
-            var result = testList.filter(obj => {
-              return obj._id === vote_id
-            })
-            var testObj = result ? result[0] : "" 
-            var test_total_votes = 0
-            test_total_votes = testObj.total_votes-1
-            testObj['total_votes'] = test_total_votes
-            testList[item] = testObj
-            set_tokens_list(testList)
-            voting_ids.splice(voting_ids.indexOf(vote_id), 1) 
-            set_voting_message(res.data.message)
-            setHandleModalVote(!handleModalVote) 
-          }
-        })
-      }
-    }
-
-    const addToWatchlist = (param_token_id) =>
-    {
-      Axios.get(API_BASE_URL+"markets/token_watchlist/add_to_watchlist/"+param_token_id, config)
-      .then(res=>
-      { 
+        const res = await Axios.get(API_BASE_URL+"markets/cryptocurrency/add_to_watchlist/"+param_token_id, config)
         if(res.data.status)
         {
-          var sdawatchlist = watchlist
-          set_watchlist([])
-          sdawatchlist.push(param_token_id)
-          set_watchlist(sdawatchlist)
+          var list = []
+          for(const i of tokens_list)
+          {  
+              if(i._id == param_token_id)
+              { 
+                  i['watchlist_status'] = true
+                  list.push(i)
+              }
+              else
+              {
+                  list.push(i)
+              }
+          }
+          set_tokens_list(list)
         }
-      })
     }
     
     const removeFromWatchlist = (param_token_id) =>
     {
-      Axios.get(API_BASE_URL+"markets/token_watchlist/remove_from_watchlist/"+param_token_id, config).then(res=>
+      Axios.get(API_BASE_URL+"markets/cryptocurrency/remove_from_watchlist/"+param_token_id, config).then(res=>
       {
         if(res.data.status)
         {
-          var sdawatchlist = watchlist
-          set_watchlist([])
-          sdawatchlist.splice(sdawatchlist.indexOf(param_token_id), 1)
-          set_watchlist(sdawatchlist)
+          var list = []
+          for(const i of tokens_list)
+          {  
+              if(i._id == param_token_id)
+              { 
+                  i['watchlist_status'] = false
+                  list.push(i)
+              }
+              else
+              {
+                  list.push(i)
+              }
+          }
+          set_tokens_list(list)
         }
       })
     }
@@ -232,25 +202,18 @@ export default function Companies({user_token, config, userAgent})
     {  
         set_watchlist_tab_status("")
         set_all_tab_status(1)
-        set_category_row_id("")
-        set_category_name("")
+        set_category_id("")
         set_search_title("")
         tokensList({selected : 0})
-    }  
+    } 
     
-    const set_tab_list=(id)=>
-    {  
-        set_business_tab_status(2)
-        set_all_tab_status(0)
-        set_category_row_id(id)
+    
+    const closeCategory = () => 
+    {
+      set_category_id("")
+      set_category_name("")
     } 
 
-    const handleChange = (selectedOption) => 
-    {
-      router.push("/category/"+selectedOption.value)
-      // set_category_row_id(selectedOption.value)
-      // set_category_name(selectedOption.label)
-    } 
 
     const makeJobSchema=()=>
     {  
@@ -263,11 +226,8 @@ export default function Companies({user_token, config, userAgent})
             "sameAs":["http://www.facebook.com/Coinpedia.org/","https://twitter.com/Coinpedianews", "http://in.linkedin.com/company/coinpedia", "http://t.me/CoinpediaMarket"]
         } 
     }
-   
-   
  
 return (
-    
    <>
       <Head>
          <title>Cryptocurrency Market Live Insights | Coinpedia </title>
@@ -295,6 +255,7 @@ return (
          <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(makeJobSchema()) }} /> 
       </Head>
 
+      {/* <div className="page new_markets_index min_height_page markets_new_design"  ref={myRef}> */}
       <div className="page new_markets_index min_height_page markets_new_design">
       <div className="market-page">
 
@@ -304,257 +265,20 @@ return (
           <div className="container">
             <div className="col-md-12">
               <div className="row market_insights ">
-                <div className="col-md-6 col-lg-6">
+                <div className="col-md-12 col-lg-6">
                   <h1 className="page_title">Cryptocurrency Live Prices</h1>
                   {/* by Market Cap */}
-                  <p>Global market cap today: $ 1.38 trillion <img src="/assets/img/markets/high-value.png" alt="high value"/><span class="color-green">10.02%</span> in 24 hrs.</p>
+                  <p>Global market cap today: $ 1.38 trillion <img src="/assets/img/markets/high-value.png" alt="high value"/><span className="color-green">10.02%</span> in 24 hrs.</p>
                 </div>
                 <div className="col-md-1 col-lg-2"></div>
                 <div className="col-md-5 col-lg-4" >
                   <SearchContractAddress /> 
                 </div>
               </div>
-              {/* <div className="row">
-                <div className="col-md-12">
-                  <ul className="trending-tokens">
-                    <li><img src="/assets/img/markets/trending.png" alt="trending"/> Trending Coin:</li>
-                    <li><img src="/assets/img/markets/bitcoin.png" alt="bitcoin"/> Bitcoin (BTC) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="color-green">03.56%</span></li>
-                    <li><img src="/assets/img/markets/arbitrum.png" alt="arbitrum"/> Arbitrum (ARB) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="color-green">04.06%</span></li>
-                    <li><img src="/assets/img/markets/shiba.png" alt="shiba"/> Shiba (INU) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="color-green">0.98%</span></li>
-                    <li><img src="/assets/img/markets/ethereum.png" alt="ethereum"/> Ethereum (ETH) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="color-green">03.84%</span></li>
-                    <li><img src="/assets/img/markets/dogecoin.png" alt="dogecoin"/> Dogecoin (DOGE) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="color-green">0.98%</span></li>
-                    <li data-toggle="modal" data-target="#trending-modal"><h6>More...</h6></li>
-                  </ul>
-
+              <div>
+                <div className="all-categories-list ">
+                  <CategoriesTab active_tab={1} user_token={user_token}/> 
                 </div>
-              </div> */}
-
-<div>
-<div class="all-categories-list">
-  <CategoriesTab /> 
-</div>
-
-
-{/* <div className="col-md-12 col-lg-6">
-  <div className="gainers-loosers">
-    <div className="row">
-      <div className="col-md-7">
-        <ul class="nav nav-tabs">
-          <li><a data-toggle="tab" href="#gainers" class="active">Gainers</a></li>
-          <li><a data-toggle="tab" href="#losers">Losers</a></li>
-        </ul>
-      </div>
-      <div className="col-md-5">
-        <a href="#" className="view-more-pg">View More <img src="/assets/img/markets/more.png" alt="view more" /></a>
-      </div>
-    </div>
-    <div className="row">
-      <div className="col-md-12">
-        <div class="tab-content">
-          <div id="gainers" class="tab-pane fade in active show">
-              <table class="table">
-                <thead>
-                  <tr>
-                    <th>Rank</th>
-                    <th>Name</th>
-                    <th>Price</th>
-                    <th>24h</th>
-                    <th class="text-right">24h Volume</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                    <td>23</td>
-                    <td><img src="/assets/img/markets/gala.png" alt="gala" /> Gala <span>GALA</span></td>
-                    <td>$ 0.043</td>
-                    <td className="text-green">7.68%</td>
-                    <td class="text-right">$156,749,287</td>
-                  </tr>
-                  <tr>
-                    <td>45</td>
-                    <td><img src="/assets/img/markets/ldo.png" alt="ldo" /> Lido DAO Token <span>LDO</span> </td>
-                    <td>$ 2.36</td>
-                    <td className="text-green">6.11%</td>
-                    <td class="text-right"> $2,388,804,583.54</td>
-                  </tr>
-                  <tr>
-                    <td>105</td>
-                    <td><img src="/assets/img/markets/nexo.png" alt="nexo" /> Nexo <span>NEXO</span></td>
-                    <td>$ 0.70</td>
-                    <td className="text-green">2.012%</td>
-                    <td class="text-right">$660,093,874.81</td>
-                  </tr>
-                  <tr>
-                    <td>10254</td>
-                    <td><img src="/assets/img/markets/stg.png" alt="stg" /> StargateToken <span>STG</span></td>
-                    <td>$ 0.83</td>
-                    <td className="text-green">6.95%</td>
-                    <td class="text-right">$881,627,054.38</td>
-                  </tr>
-                </tbody>
-              </table>
-              <a href={app_coinpedia_url+"gainers-and-losers/?active_category=3"} className="view-more-pg">View More <img src="/assets/img/markets/more.png" alt="view more" /></a>
-          </div>
-          <div id="losers" class="tab-pane fade">
-          <div id="gainers" class="tab-pane fade in active show">
-              <table class="table">
-                <thead>
-                  <tr>
-                    <th>Rank</th>
-                    <th>Name</th>
-                    <th>Price</th>
-                    <th>24h</th>
-                    <th class="text-right">24h Volume</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                    <td>23</td>
-                    <td><img src="/assets/img/markets/gala.png" alt="gala" /> Gala <span>GALA</span></td>
-                    <td>$ 0.043</td>
-                    <td className="text-red">7.68%</td>
-                    <td class="text-right">$156,749,287</td>
-                  </tr>
-                  <tr>
-                    <td>45</td>
-                    <td><img src="/assets/img/markets/ldo.png" alt="ldo" /> Lido DAO Token <span>LDO</span> </td>
-                    <td>$ 2.36</td>
-                    <td className="text-red">6.11%</td>
-                    <td class="text-right">$2,388,804,583.54</td>
-                  </tr>
-                  <tr>
-                    <td>105</td>
-                    <td><img src="/assets/img/markets/nexo.png" alt="nexo" /> Nexo <span>NEXO</span></td>
-                    <td>$ 0.70</td>
-                    <td className="text-red">2.012%</td>
-                    <td class="text-right">$660,093,874.81</td>
-                  </tr>
-                  <tr>
-                    <td>10254</td>
-                    <td><img src="/assets/img/markets/stg.png" alt="stg" /> StargateToken <span>STG</span></td>
-                    <td>$ 0.83</td>
-                    <td className="text-red">6.95%</td>
-                    <td class="text-right">$881,627,054.38</td>
-                  </tr>
-                </tbody>
-              </table>
-          </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div> */}
-
-
-{/* <div className="col-md-12 col-lg-6">
-<div className="gainers-loosers top-categories">
-    <div className="row">
-      <div className="col-md-12">
-          <h2>Top Categories</h2>
-      </div>
-    </div>
-    <div className="row">
-      <div className="col-md-12">
-        <div class="tab-pane">
-              <table class="table">
-                <thead>
-                  <tr>
-                  
-                    <th>Name</th>
-                    <th>Dominance</th>
-                    <th className="text-right">Gainer / Looser Number</th>
-                  
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                    <td> DeFi</td>
-                    <td className="text-green">11.77%</td>  
-                    <td>
-                        <table>
-                          <tbody>
-                              <tr>
-                                <td><span>147(79%)</span></td>
-                                <td className="gain-lose-progress">
-                                  <div class="progress">
-                                      <div class="progress-bar" style={{width:'79%'}}></div>
-                                  </div>
-                                </td>
-                                <td><span>38(21%)</span></td>
-                            </tr>
-                          </tbody>
-                        </table>
-                      </td>
-                  </tr>
-                  <tr>
-                    <td> Jobs</td>
-                    <td className="text-green">56.23%</td>
-                    <td>
-                        <table>
-                          <tbody>
-                              <tr>
-                                <td><span>3(100%)</span></td>
-                                <td className="gain-lose-progress">
-                                  <div class="progress">
-                                      <div class="progress-bar" style={{width:'100%'}}></div>
-                                  </div>
-                                </td>
-                                <td><span>0(0%)</span></td>
-                            </tr>
-                          </tbody>
-                        </table>
-                      </td>
-                  </tr>
-                  <tr>
-                    <td> BNB Chain Ecosystem</td>
-                    <td className="text-green">16.96%</td>
-                    <td>
-                        <table>
-                          <tbody>
-                              <tr>
-                                <td><span>147(60%)</span></td>
-                                <td className="gain-lose-progress">
-                                  <div class="progress">
-                                      <div class="progress-bar" style={{width:'60%'}}></div>
-                                  </div>
-                                </td>
-                                <td><span>68(40%)</span></td>
-                            </tr>
-                          </tbody>
-                        </table>
-                      </td>
-                  </tr>
-                  <tr>
-                    <td> NFT</td>
-                    <td className="text-green">0.55%</td>
-                    <td>
-                        <table>
-                          <tbody>
-                              <tr>
-                                <td><span>147(59%)</span></td>
-                                <td className="gain-lose-progress">
-                                  <div class="progress">
-                                      <div class="progress-bar" style={{width:'59%'}}></div>
-                                  </div>
-                                </td>
-                                <td><span>38(41%)</span></td>
-                            </tr>
-                          </tbody>
-                        </table>
-                      </td>
-                  </tr>
-                </tbody>
-              </table>
-              <a href="#" className="view-more-pg">View More <img src="/assets/img/markets/more.png" alt="view more" /></a>
-          </div>
-      </div>
-    </div>
-  </div>
-</div> */}
-
-
-
               </div>
             </div>
           </div>
@@ -563,93 +287,94 @@ return (
         
           <div className="container price-tracking-tbl">
           <div className="col-md-12">
-          {/* <div className="row"> */}
-          {/* <div className="col-md-12">
-            {
-              <CategoriesTab /> 
-            }
-            </div> */}
-            {/* <div className="col-md-12">
-              <h2>Price Tracking</h2>
-            </div> */}
-            {/* <div className="col-md-3"> */}
-              {/* <div className="input-group search_filter">
-                <input value={search_title} onChange={(e)=> set_search_title(e.target.value)} type="text" className="form-control search-input-box" placeholder="Search Token By Name" />
-                  <div className="input-group-prepend ">
-                    <span className="input-group-text" onClick={()=> tokensList({selected:0})}><img src="/assets/img/search_large.svg" alt="search-box"  width="100%" height="100%"/></span>                 
-                  </div>
-              </div>  */}
-              {/* <div className="error">  {err_searchBy}</div>
-              {validSearchContract && <div className="error">{validSearchContract}</div>} */}
-            {/* </div>  */}
-          {/* </div> */}
-            
-          
-            
-              {/* <div className="">
-              <div className="row">
-                <div className="col-lg-4 col-sm-6 col-12">
-                  <p className="companies_found">{total_tokens_count} Tokens Found</p>
-                </div>
-              </div>
-              </div> */}
-
-
-
               <div className="prices transaction_table_block">
                 <div className="row">
                   <div className="col-md-12 col-12">
-                  <div className="row">
-                         <div className="col-md-12 col-lg-8 col-12">
-                          <ul className="category_list">
-                            {
-                              user_token?
-                              <li className={watchlist_tab_status===2?"active_tab":null}><Link href={app_coinpedia_url+"watchlist/?active_watchlist_tab=2"}><a><img src="/assets/img/wishlist_star.svg" alt="Watchlist"/> Watchlist</a></Link></li>
-                              :
-                              <li>
-                              <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url}><a onClick={()=> Logout()}><img src="/assets/img/wishlist_star.svg" alt="Watchlist" width={17} height={17} /> Watchlist</a></Link>
-                              </li>
-                            }
-                            
-                            <li>
-                              <Link href={app_coinpedia_url}><a><img src="https://image.coinpedia.org/wp-content/uploads/2022/06/21160228/menu-wallet.svg" alt="Portfolio" width={20} height={20} /> Portfolio</a></Link>
-                            </li>
-                            <li>|</li>
+                    <div className="row">
+                      <div className="col-md-12 col-lg-5 col-12">
+                        <h4 className="markets_subtitle">Price Tracking</h4>
+                      </div>
+                      <div className="col-md-12 col-lg-7 col-12  filter-category-section">
+                        <div className="row">
+                          <div className="col-md-5 col-6 col-sm-6">
+                            <div className="cust_filters_dropdown" ref={category_ref}>
+                              <div className="cust_filter_input" >
+                               
+                                {
+                                  !category_id ?
+                                  <div className="input-group" onClick={() =>
+                                      set_category_status(!category_status)
+                                    }>
 
-                            <li className={all_tab_status===1?"active_tab":null}><a onClick={()=>set_all_tab_active()}>All</a></li>
+                                      <input onChange={(e)=> searchTags(e.target.value)} autoComplete='off' type="text" className="form-control " placeholder="Select Category" />
+                                      <span className="input-group-addon lightmode_image">
+                                        <img src="/assets/img/filter_dropdown.svg" title="Filter Dropdown" alt="Filter Dropdown" />
+                                      </span>
+                                      <span className="input-group-addon darkmode_image">
+                                        <img src="/assets/img/filter_dropdown_grey.svg" title="Filter Dropdown" alt="Filter Dropdown" />
+                                      </span>
+                                  </div>
+                                  :
+                                  <div className="markets_selected_category">
+                                    {category_name}
+
+                                    <span className="input-group-addon close_category_icon" onClick={()=>closeCategory()}>
+                                      <img src="/assets/img/close_mark.png" />
+                                    </span>
+                                  </div>  
+                                }
+                              </div>
+                              {
+                                category_status ?
+                                <ul className="cust_filter_result">
+                                  {
+                                    category_list.length ?
+                                    category_list.map((item, i) =>
+                                      <li onClick={()=>{
+                                        set_category_id(item._id)
+                                        set_category_name(item.category_name)
+                                        set_category_status(false)
+                                      }
+                                    }> {item.category_name} </li>
+                                    )
+                                    :
+                                    ""
+                                  }
+                                </ul>
+                                :
+                                ""
+                              }
+                              
+                            </div>
                             
-                            {
-                              category_list.map((e,i)=>
-                              // <li class={parseInt(category_row_id)===parseInt(e.category_row_id) && business_tab_status===2?"active_tab":null}><a onClick={()=>set_tab_list(e.category_row_id)}>{e.category_name}</a></li>
-                              <li>
-                              <Link href={market_coinpedia_url+"category/"+e.category_id}><a>{e.category_name}</a></Link>
-                              </li>
-                              )
-                           }
-                          </ul>
-                        </div>
-                       <div className="col-md-12 col-lg-4 col-12 filter-category-section">
-                  
-                      <div className="row">
-                        <div className="col-md-9 col-12 col-sm-6">
-                          <div className="input-group search_filter">
-                            <input value={search_title} onChange={(e)=> set_search_title(e.target.value)} type="text" className="form-control search-input-box" placeholder="Search Token By Name" />
+
+
+                            {/* <Select
+                            onChange={handleChange}
+                            options={other_category_list}
+                            placeholder={category_name?category_name:'Select  Category'}
+                            value={category_name}
+                            />  */}
+                          </div>
+                      
+                          <div className="col-md-5 col-6 col-sm-6 ">
+                            <div className="input-group search_filter">
+                              <input value={search_title} onChange={(e)=> set_search_title(e.target.value)} type="text" className="form-control search-input-box" placeholder="Search Token By Name" />
                               <div className="input-group-prepend ">
-                                  <span className="input-group-text" onClick={()=> tokensList({selected:0})}><img src="/assets/img/search_large.svg" alt="search-box"  width="100%" height="100%"/></span>                 
-                                </div>
-                          </div> 
-                         
+                                <span className="input-group-text" onClick={()=> tokensList({selected:0})}><img src="/assets/img/search_large.svg" alt="search-box"  width="100%" height="100%"/></span>                 
+                              </div>
+                            </div> 
                           </div>
 
-                          <div className="col-md-3 col-6 ">
+                          <div className="col-md-2 col-6 mobile_hide_view">
                             <ul className="filter_rows">
-                                <li>{count>100?
+                                <li>
                                   <select className="show-num" onChange={(e)=>set_per_page_count(e.target.value)} >
-                                  <option value="" disabled>Show Rows</option>
-                                  <option value={100}>100</option>
-                                  <option value={50}>50</option>
-                                  <option value={20}>20</option>
-                                  </select>:null}
+                                    <option value="" disabled>Show Rows</option>
+                                    <option value={100}>100</option>
+                                    <option value={50}>50</option>
+                                    <option value={20}>20</option>
+                                  </select>
                                 </li>
                             </ul>
                           </div>
@@ -665,33 +390,17 @@ return (
                        <table className="table table-borderless">
                          <thead>
                             <tr>
-
-
-                                <th className="" style={{minWidth: '34px'}}></th>
-                                <th className="mobile_hide_table_col" style={{minWidth: '34px'}}>#</th>
-                                <th className="">Name</th>
+                                <th className="mobile_fixed_first" style={{minWidth: '35px'}}></th>
+                                <th className="mobile_hide_view" style={{minWidth: '35px'}}>#</th>
+                                <th className="mobile_fixed rank_class">Name</th>
                                 <th className="">Price</th>
-                                <th className=" mobile_hide_table_col" style={{minWidth: 'unset'}}>1h</th>
-                                <th className="" style={{minWidth: 'unset'}}>24h</th>
-                                <th className=" mobile_hide_table_col" style={{minWidth: 'unset'}}>7d</th>
-                                <th className="mobile_hide_table_col table_circulating_supply">Market Cap</th> 
-                                <th className=" mobile_hide_table_col">Volume(24H)</th>  
-                                <th className="mobile_hide_table_col table_circulating_supply">Circulating Supply</th>  
-                                <th className="mobile_hide_table_col">Last 7 Days</th>
-
-                                {/* <th className="" style={{minWidth: '34px'}}></th> */}
-                                {/* <th className="">Rank</th>
-                                <th className="">Name</th>
-                                <th className="table_token">Price</th>
-                                <th className="table_token mobile_hide_table_col">Type</th>
-                                <th className="table_max_supply mobile_hide_table_col">Max Supply</th> 
-                                <th className="table_token mobile_hide_table_col">1 H %</th>
-                                <th className="table_token mobile_hide_table_col">24 H %</th>
-                                <th className="table_token mobile_hide_table_col">7 D %</th>
-                                <th className="mobile_hide_table_col table_circulating_supply">Market Cap</th>  
-                                <th className="mobile_hide_table_col table_circulating_supply">Circulating Supply</th>  
-                                <th className="table_circulating_supply mobile_hide_table_col text-center">Votes</th>  
-                                <th className="table_circulating_supply mobile_hide_table_col text-center">Action</th>   */}
+                                <th className="  " style={{minWidth: 'unset'}}>1h</th>
+                                <th className="  " style={{minWidth: 'unset'}}>24h</th>
+                                <th className="  " style={{minWidth: 'unset'}}>7d</th>
+                                <th className="  table_circulating_supply">Market Cap</th> 
+                                <th className="  ">Volume(24H)</th>  
+                                <th className="  table_circulating_supply">Circulating Supply</th>  
+                                <th className=" ">Last 7 Days</th>
                             </tr>
                          </thead>
                          
@@ -701,109 +410,56 @@ return (
                             loader_status ?
                            <>
                            {
-                             tokens_list.length > 0
-                             ?
+                             tokens_list.length > 0 ?
                              tokens_list.map((e, i) => 
                              <tr key={i}>
                                  
-                                     <td>
-                                     
-                                     {
-                                       user_token ?
-                                       <>
-                                       {
-                                         watchlist.includes(e._id) ?
-                                        
-                                        //  /assets/img/markets/filled-star.svg
-                                        // /assets/img/markets/star.svg
-                                         <span onClick={()=>removeFromWatchlist(e._id)} ><img src=" /assets/img/wishlist_star_selected.svg" alt="Watchlist" width={17} height={17} /></span>
-                                         :
-                                         <span onClick={()=>addToWatchlist(e._id)} ><img src="/assets/img/wishlist_star.svg" alt="Watchlist" width={17} height={17} /></span>
-                                         }
-                                       </>
-                                       :
-                                       <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url}><a onClick={()=> Logout()}><img src="/assets/img/wishlist_star.svg" alt="Watchlist" width={17} height={17} /></a></Link>
-                                     }
-                                     
+                                     <td className="mobile_fixed_first">
+                                      {
+                                        user_token ?
+                                          <>
+                                          {
+                                            e.watchlist_status== true ?
+                                            <span onClick={()=>removeFromWatchlist(e._id)} ><img src=" /assets/img/color.svg" alt="Watchlist" width={17} height={17} /></span>
+                                            :
+                                            <span onClick={()=>addToWatchlist(e._id)} ><img src="/assets/img/star.svg" alt="Watchlist" width={17} height={17} /></span>
+                                            }
+                                          </>
+                                          :
+                                          <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url} onClick={()=> Logout()}><img src="/assets/img/star.svg" alt="Watchlist"/></Link>
+                                      }
                                      </td>
-                                    <td>
-                                        <span className="mobile_hide_table_col wishlist"> {sl_no+i+1}</span>
+                                    <td className="mobile_hide_view mobile_td_fixed wishlist"> {sl_no+i+1}
                                     </td>
-                                     <td>
+                                     <td className="mobile_fixed rank_class">
                                        <Link href={"/"+e.token_id}>
-                                         <a>
+                                         
                                           <div className="media">
-                                            <div className="media-left">
+                                            <div className="media-left align-self-center">
                                               <img src={image_base_url+(e.coinmarketcap_id ? e.coinmarketcap_id+".png" : "default.png")} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt={e.token_name} width="100%" height="100%" className="media-object" />
                                             </div>
-                                            <div className="media-body">
-                                              <h4 className="media-heading">{e.token_name} <span>{(e.symbol).toUpperCase()}</span></h4>
+                                            <div className="media-body align-self-center">
+                                              <h4 className="media-heading">{strLenTrim(e.token_name, 14)} </h4>
+                                              <p>{(e.symbol).toUpperCase()}</p>
                                             </div>
                                           </div> 
-                                         </a>
+                                         
                                        </Link>
                                      </td> 
-                                     {/* <td>{e.price === null? "-":"$"+ Number(e.price).toFixed(2)}</td> */}
-                                     {/* <td>
-                                       <span className="twenty_high"><img src="/assets/img/green-up.png" />2.79%</span>
-                                     </td> */}
-               
                                      <td className="market_list_price"> 
                                        <Link href={"/"+e.token_id}>
-                                         <a>
+                                         
                                          <span className="block_price">{roundNumericValue(e.price)}</span>
-                                           {/* <span className="block_price">{e.price?"$":null}{e.price?parseFloat((e.price).toFixed(9)):"-"}</span> */}
-                                          {/* <br/> {e.price} <br/> */}
+                                          
                                            {e.updated_on ? moment(e.updated_on).fromNow():null} 
-                                         </a>
+                                         
                                          </Link>
                                     </td>
                                     
-                                    {/* <td className="mobile_hide_table_col"> 
-                                       <Link href={"/"+e.token_id}>
-                                        <a>
-                                          {
-                                            e.list_type==1?
-                                            "Coin"
-                                            :
-                                            (e.contract_addresses) && ((e.contract_addresses).length ===1)
-                                            ?
-                                            parseInt(e.contract_addresses[0].network_type) === 1 ? "ERC20" : "BEP20" 
-                                            // e.contract_addresses.map((ca)=>
-                                            //   parseInt(ca.network_type) === 1 ? "ERC20" : "BEP20" 
-                                            //)
-                                            :
-                                            (e.contract_addresses) && (e.contract_addresses).length === 2
-                                            ?
-                                            e.contract_addresses.map((ca,i)=>
-                                              i==0?
-                                              parseInt(ca.network_type) === 1? "ERC20"+","
-                                              :
-                                              "BEP20" +","
-                                              :
-                                              parseInt(ca.network_type) === 1? "ERC20"
-                                              :
-                                              "BEP20" 
-                                            )
-                                            :
-                                            null
-                                            
-                                          } 
-                                        </a>
-               
-                                         </Link>
-                                    </td> */}
                                     
-                                    {/* <td className="mobile_hide_table_col">
-                                       <Link href={"/"+e.token_id}>
-                                         <a>
-                                           {e.total_max_supply ? separator(e.total_max_supply) : "-"} 
-                                         </a>
-                                       </Link>
-                                    </td> */}
 
-                                    <td className="mobile_hide_table_col">
-                                       <Link href={"/"+e.token_id}><a>
+                                    <td className=" ">
+                                       <Link href={"/"+e.token_id}>
                                          {
                                         e.percent_change_1h?e.percent_change_1h>0?
                                         <h6 className="values_growth"><span className="green"><img src="/assets/img/markets/high.png" alt="high price"/>{e.percent_change_1h.toFixed(2)+"%"}</span></h6>
@@ -813,11 +469,11 @@ return (
                                         "--"
                                         }
                                         {/* <h6 className="values_growth"><span className="red"><img src="/assets/img/markets/low.png" alt="high price"/>1.05%</span></h6> */}
-                                       </a></Link>
+                                       </Link>
                                      </td>
                
-                                    <td className="mobile_hide_table_col">
-                                       <Link href={"/"+e.token_id}><a>
+                                    <td className=" ">
+                                       <Link href={"/"+e.token_id}>
                                        {
                                         e.percent_change_24h?e.percent_change_24h>0?
                                         <h6 className="values_growth"><span className="green"><img src="/assets/img/markets/high.png" alt="high price"/>{e.percent_change_24h.toFixed(2)+"%"}</span></h6>
@@ -826,11 +482,11 @@ return (
                                         :
                                         "--"
                                         }
-                                       </a></Link>
+                                       </Link>
                                      </td>
 
-                                     <td className="mobile_hide_table_col">
-                                       <Link href={"/"+e.token_id}><a>
+                                     <td className=" ">
+                                       <Link href={"/"+e.token_id}>
                                        {
                                         e.percent_change_7d?e.percent_change_7d>0?
                                         <h6 className="values_growth"><span className="green"><img src="/assets/img/markets/high.png" alt="high price"/>{e.percent_change_7d.toFixed(2)+"%"}</span></h6>
@@ -840,61 +496,34 @@ return (
                                         "--"
                                         }
                                         {/* <h6 className="values_growth"><span className="red"><img src="/assets/img/markets/low.png" alt="high price"/>1.05%</span></h6> */}
-                                       </a></Link>
+                                       </Link>
                                      </td>
 
-                                     <td className="mobile_hide_table_col">
-                                       <Link href={"/"+e.token_id}><a>
-                                       {e.circulating_supply ?"$"+separator((e.circulating_supply*e.price).toFixed(0)) : "-"}
-                                       </a></Link>
-                                     </td>  
+                                     <td className=" ">
+                                        <Link href={"/"+e.token_id}>
+                                        {e.circulating_supply ?"$"+separator((e.circulating_supply*e.price).toFixed(0)) : "-"}
+                                        </Link>
+                                    </td>  
                                      
-                                     <td className="mobile_hide_table_col">
-                                       <Link href={"/"+e.token_id}><a>
+                                     <td className=" ">
+                                       <Link href={"/"+e.token_id}>
                                        {e.volume ?"$"+separator((e.volume).toFixed(0)) : "-"}
-                                       </a></Link>
+                                       </Link>
                                      </td>
 
-                                     <td className="mobile_hide_table_col">
-                                       <Link href={"/"+e.token_id}><a>
+                                     <td className=" ">
+                                     <div className='circulating-supply'>
+                                      <Link href={"/"+e.token_id}>
                                        {e.circulating_supply ? separator((e.circulating_supply).toFixed(0))+" "+(e.symbol).toUpperCase() : "-"}
-                                       </a></Link>
-                                     </td>
+                                       </Link>
+                                      </div>
+                                     </td> 
 
                                      
                                       
-                                     <td className="mobile_hide_table_col">
-                                      <img className={e.percent_change_7d>0 ? "saturated-up":"saturated-down"} src={"https://s3.coinmarketcap.com/generated/sparklines/web/7d/2781/"+(e.coinmarketcap_id ? e.coinmarketcap_id+".svg":"")} onError={(e) =>e.target.src = ""} alt={e.token_name} width="100%" height="100%"/>
-                                     </td> 
-                                       {/* <td  className="mobile_hide_table_col text-center">
-                                       {
-                                       e.total_votes == 0 ?
-                                       voting_ids.includes(e._id) ? <span className="vote_value">1</span> : "--"
-                                       :
-                                       <Link href={"/"+e.token_id}>
-                                         <a>
-                                           <span className="vote_value">{e.total_votes}</span>
-                                         </a>
-                                       </Link>
-                                       }
-                                     </td> */}
-                                     
-                                     
-                                     {/* <td  className="mobile_hide_table_col text-center">
-                                       {
-                                         user_token ?
-                                         <>
-                                         {
-                                           voting_ids.includes(e._id) ?
-                                           <span className="market_list_price markets_voted"> <button data-toggle="tooltip" onClick={()=>ModalVote(e.token_id,true,e._id,i)} >Voted</button></span>
-                                           :
-                                           <span className="market_list_price"><button data-toggle="tooltip" onClick={()=>ModalVote(e.token_id,false,e._id,i)} className="vote_btn">Vote</button></span>
-                                          }
-                                         </>
-                                         :
-                                         <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url}><a onClick={()=> Logout()}><span className="market_list_price"><button data-toggle="tooltip" className="vote_btn">Vote</button></span></a></Link>
-                                       }
-                                       </td>  */}
+                                     <td className=" ">
+                                      <img className={e.percent_change_7d>0 ? "saturated-up":"saturated-down"} src={"https://s3.coinmarketcap.com/generated/sparklines/web/7d/2781/"+(e.coinmarketcap_id ? e.coinmarketcap_id+".svg":"")} onError={(e) =>e.target.src = ""} alt={e.token_name} style={{width:"170px"}} height="100%"/>
+                                     </td>
                                </tr> 
                              ) 
                              :
@@ -906,7 +535,7 @@ return (
                            }
                              </>
                              :
-                             <TableContentLoader row="10" col="10" />  
+                             <TableContentLoader row="10" col="11" />  
                           }
                            
                          </tbody>
@@ -977,66 +606,32 @@ return (
             </div>
           </div>
       </div>
-      <div className={"modal connect_wallet_error_block"+ (handleModalVote ? " collapse show" : "")}>
-            <div className="modal-dialog modal-sm">
-              <div className="modal-content">
-                <div className="modal-body">
-                  <button type="button" className="close" data-dismiss="modal" onClick={()=>ModalVote()}>&times;</button>
-                  {
-                    voting_status == false ?
-                    <h4> Do you want to support this token ? </h4>
-                    :
-                    <h4> Do not support this token ? </h4>
-                  }
-                  
-                  <div className="vote_yes_no">
-                    {/* className={voting_status == 1 ? "vote_yes" : "vote_no"} */}
-                    {
-                      voting_status == false ?
-                      <>
-                      <button onClick={()=>vote(1)}>Confirm</button>  
-                      <button onClick={()=>ModalVote()}>Cancel</button>
-                      </>
-                      :
-                      <>
-                      <button onClick={()=>vote(0)}>Confirm</button>  
-                      <button onClick={()=>ModalVote()}>Cancel</button>
-                      </>
-                    }
-                    
-                  </div>
-                </div>
-              </div> 
-            </div>
-          </div> 
 
-          <div class="modal" id="trending-modal">
-            <div class="modal-dialog modal-lg">
-              <div class="modal-content">
+        <div className="modal" id="trending-modal">
+          <div className="modal-dialog modal-lg">
+            <div className="modal-content">
+            
+              <div className="modal-header">
+            
+                <button type="button" className="close" data-dismiss="modal">&times;</button>
+              </div>
               
-                <div class="modal-header">
-              
-                  <button type="button" class="close" data-dismiss="modal">&times;</button>
-                </div>
-                
-                <div class="modal-body">
-          
-                  <ul className="trending-tokens">
-                    <li><img src="/assets/img/markets/trending.png" alt="trending"/> Trending Coin:</li>
-                    <li><img src="/assets/img/markets/bitcoin.png" alt="bitcoin"/> Bitcoin (BTC) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">03.56%</span></li>
-                    <li><img src="/assets/img/markets/arbitrum.png" alt="arbitrum"/> Arbitrum (ARB) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">04.06%</span></li>
-                    <li><img src="/assets/img/markets/shiba.png" alt="shiba"/> Shiba (INU) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">0.98%</span></li>
-                    <li><img src="/assets/img/markets/ethereum.png" alt="ethereum"/> Ethereum (ETH) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">03.84%</span></li>
-                    <li><img src="/assets/img/markets/dogecoin.png" alt="dogecoin"/> Dogecoin (DOGE) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">0.98%</span></li>
-                    <li><img src="/assets/img/markets/dogecoin.png" alt="dogecoin"/> Dogecoin (DOGE) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">0.98%</span></li>
-                  </ul>
+              <div className="modal-body">
+                <ul className="trending-tokens">
+                  <li><img src="/assets/img/markets/trending.png" alt="trending"/> Trending Coin:</li>
+                  <li><img src="/assets/img/markets/bitcoin.png" alt="bitcoin"/> Bitcoin (BTC) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">03.56%</span></li>
+                  <li><img src="/assets/img/markets/arbitrum.png" alt="arbitrum"/> Arbitrum (ARB) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">04.06%</span></li>
+                  <li><img src="/assets/img/markets/shiba.png" alt="shiba"/> Shiba (INU) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">0.98%</span></li>
+                  <li><img src="/assets/img/markets/ethereum.png" alt="ethereum"/> Ethereum (ETH) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">03.84%</span></li>
+                  <li><img src="/assets/img/markets/dogecoin.png" alt="dogecoin"/> Dogecoin (DOGE) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">0.98%</span></li>
+                  <li><img src="/assets/img/markets/dogecoin.png" alt="dogecoin"/> Dogecoin (DOGE) <span><img src="/assets/img/markets/up.png" alt="up" /></span><span className="text-green">0.98%</span></li>
+                </ul>
 
-                </div>
-          
-                
               </div>
             </div>
           </div>
+        </div>
+
     </div>
 </>
 )
