@@ -151,21 +151,21 @@ export default function OngoingLaunchPad({userAgent}) {
           <div className=" launchpad">
             <div className="container">
               <div className="col-md-12">
-              <div className="breadcrumb_block">
-              <Link href={coinpedia_url}><a >Home</a></Link> <span> &#62; </span> 
-              <Link href={market_coinpedia_url}><a >Live Market</a></Link><span> &#62; </span> Ongoing Launchpads
-               </div>
+              {/* <div className="breadcrumb_block">
+              <Link href={coinpedia_url}>Home</Link> <span> &#62; </span> 
+              <Link href={market_coinpedia_url}>Live Market</Link><span> &#62; </span> Ongoing Launchpads
+               </div> */}
                 <div className="row launchpad_list">
-                  <div className="col-md-8 col-7">
+                  <div className="col-md-8 col-12">
                     <h1 className="page_title">Now Live ! Ongoing Crypto Launchpads List</h1>
                     <p>All current and ongoing launchpads, listed in date order.</p>
                   </div>
-                  <div className="col-md-4 col-5">
+                  <div className="col-md-4 col-12">
                     <div className="launchpad-toke-button">
                     <Link href={user_token ? "/token/create-new":app_coinpedia_url+"login"}>
-                      <a><button className="btn-gradient-primary">
+                     <button className="btn_create button_transition">
                         <span className="btn__text">List your Token</span>
-                      </button></a></Link>
+                      </button></Link>
                     </div>
                   </div> 
                 </div>
@@ -173,10 +173,10 @@ export default function OngoingLaunchPad({userAgent}) {
                 <div className="row">
                   <div className="tabs_for_navigation">
                     <ul className="nav nav-tabs">
-                      <li><Link href="/launchpad"><a><span>Overview</span></a></Link></li>
-                      <li><Link href="/launchpad/ongoing"><a className="active"><span>Ongoing</span></a></Link></li>
-                      <li><Link href="/launchpad/upcoming"><a><span>Upcoming</span></a></Link></li>
-                      <li><Link href="/launchpad/ended"><a><span>Completed</span></a></Link></li>
+                      <li><Link href="/launchpad"> <span>Overview</span> </Link></li>
+                      <li><Link href="/launchpad/ongoing" className="active"><span >Ongoing</span> </Link></li>
+                      <li><Link href="/launchpad/upcoming"> <span>Upcoming</span> </Link></li>
+                      <li><Link href="/launchpad/ended"> <span>Completed</span></Link></li>
                     </ul>
                   </div>
                 </div> 
@@ -188,10 +188,10 @@ export default function OngoingLaunchPad({userAgent}) {
                         <li className="active_tab">All</li>
                         {
                               user_token?
-                              <li className={watchlist_tab_status===2?"active_tab":null}><Link href={app_coinpedia_url+"watchlist/?active_watchlist_tab=2"}><a><img src="/assets/img2/Star.svg" alt="Watchlist"/> Watchlist</a></Link></li>
+                              <li className={watchlist_tab_status===2?"active_tab":null}><Link href="/watchlist"><img src="/assets/img/star.svg" alt="Watchlist"/> Watchlist</Link></li>
                               :
                               <li>
-                              <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url+"launchpad/ongoing"}><a onClick={()=> Logout()}><img src="/assets/img2/Star.svg" alt="Watchlist"/> Watchlist</a></Link>
+                              <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url+"launchpad/ongoing"} onClick={()=> Logout()}><img src="/assets/img/star.svg" alt="Watchlist"/> Watchlist</Link>
                               </li>
                             }
                         <li className="inactive"  data-toggle="modal" data-target="#comingSoon">DeFi</li>
@@ -247,13 +247,13 @@ export default function OngoingLaunchPad({userAgent}) {
                             <>
                             {
                               watchlist.includes(e.token_row_id) ?
-                              <span onClick={()=>removeFromWatchlist(e.token_row_id)} ><img src="/assets/img2/Color.svg" alt="Watchlist"/></span>
+                              <span onClick={()=>removeFromWatchlist(e.token_row_id)} ><img src="/assets/img/color.svg" alt="Watchlist"/></span>
                               :
-                              <span onClick={()=>addToWatchlist(e.token_row_id)} ><img src="/assets/img2/Star.svg" alt="Watchlist"/></span>
+                              <span onClick={()=>addToWatchlist(e.token_row_id)} ><img src="/assets/img/star.svg" alt="Watchlist"/></span>
                               }
                             </>
                             :
-                            <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url+"launchpad/ongoing"}><a onClick={()=> Logout()}><img src="/assets/img2/Star.svg" alt="Watchlist"/></a></Link>
+                            <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url+"launchpad/ongoing"} onClick={()=> Logout()}><img src="/assets/img/star.svg" alt="Watchlist"/></Link>
                           }
                           </td>
                           
@@ -261,8 +261,10 @@ export default function OngoingLaunchPad({userAgent}) {
                           <td>
                             <a href={"/"+e.token_id}>
                               <div className="media">
+                              <div className='media-left align-self-center'>
                                 <img src={image_base_url+(e.token_image ? e.token_image : "default.png")} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt="Logo" />
-                                <div className="media-body">
+                                </div>
+                                <div className="media-body align-self-center">
                                   <h5 className="launchpad_token_title">{e.token_name} <span>{e.symbol}</span></h5>
                                 </div>
                               </div>
