@@ -150,23 +150,23 @@ export default function UpcomingLaunchPad({userAgent})
           <div className=" launchpad">
               <div className="container">
               <div className="col-md-12">
-              <div className="breadcrumb_block">
-              <Link href={coinpedia_url}><a >Home</a></Link> <span> &#62; </span> 
-              <Link href={market_coinpedia_url}><a >Live Market</a></Link><span> &#62; </span> Upcoming Launchpads
-               </div> 
+              {/* <div className="breadcrumb_block">
+              <Link href={coinpedia_url}>Home </Link> <span> &#62; </span> 
+              <Link href={market_coinpedia_url}>Live Market </Link><span> &#62; </span> Upcoming Launchpads
+               </div>  */}
               <div className="row launchpad_list">
-                <div className="col-md-8 col-7">
+                <div className="col-md-8 col-12">
                   <h1 className="page_title">Upcoming launchpad List</h1>
                   <p>Here’s the list of the top upcoming  crypto launchpads in 2022 you need to consider.</p>
                 </div>
-                <div className="col-md-4 col-5">
+                <div className="col-md-4 col-12">
                       <div className="launchpad-toke-button">
                         {/* <Link href="/launchpad/create-token"><a><button className="wallets__btn btn_blue">
                           <span className="btn__text"><img src="/assets/img/create-token-icon.png" />Create token</span>
                         </button></a></Link> */}
-                        <Link href={user_token ? "/token/create-new":app_coinpedia_url+"login" }><a><button className="btn-gradient-primary">
+                        <Link href={user_token ? "/token/create-new":app_coinpedia_url+"login" }><button className="btn_create button_transition">
                           <span className="btn__text">List your Token</span>
-                        </button></a></Link>
+                        </button> </Link>
                       </div>
                 </div> 
               </div>
@@ -174,10 +174,10 @@ export default function UpcomingLaunchPad({userAgent})
               <div className="row">
                 <div className="tabs_for_navigation">
                   <ul className="nav nav-tabs">
-                    <li><Link href="/launchpad"><a><span>Overview</span></a></Link></li>
-                    <li><Link href="/launchpad/ongoing"><a><span>Ongoing</span></a></Link></li>
-                    <li><Link href="/launchpad/upcoming"><a className="active"><span>Upcoming</span></a></Link></li>
-                    <li><Link href="/launchpad/ended"><a><span>Completed</span></a></Link></li>
+                    <li><Link href="/launchpad"> <span>Overview</span> </Link></li>
+                    <li><Link href="/launchpad/ongoing"> <span>Ongoing</span> </Link></li>
+                    <li><Link href="/launchpad/upcoming" className="active"><span>Upcoming</span> </Link></li>
+                    <li><Link href="/launchpad/ended"> <span>Completed</span> </Link></li>
                   </ul>
                 </div>
               </div>
@@ -191,10 +191,10 @@ export default function UpcomingLaunchPad({userAgent})
                       <li className="active_tab">All</li>
                       {
                               user_token?
-                              <li><Link href={app_coinpedia_url+"watchlist/?active_watchlist_tab=2"}><a><img src="/assets/img2/Star.svg" alt="Watchlist"/> Watchlist</a></Link></li>
+                              <li><Link href="/watchlist"> <img src="/assets/img/star.svg" alt="Watchlist"/> Watchlist </Link></li>
                               :
                               <li>
-                              <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url+"launchpad/upcoming"}><a onClick={()=> Logout()}><img src="/assets/img2/Star.svg" alt="Watchlist"/> Watchlist</a></Link>
+                              <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url+"launchpad/upcoming"} onClick={()=> Logout()}><img src="/assets/img/star.svg" alt="Watchlist"/> Watchlist </Link>
                               </li>
                             }
                       {/* <li><Link href={app_coinpedia_url+"?active_watchlist_tab=2"}><a><img src="/assets/img/wishlist_star.svg" alt="Watchlist"/> Watchlist</a></Link></li> */}
@@ -254,21 +254,23 @@ export default function UpcomingLaunchPad({userAgent})
                             <>
                             {
                               watchlist.includes(e.token_row_id) ?
-                              <span onClick={()=>removeFromWatchlist(e.token_row_id)} ><img src="/assets/img2/Color.svg" alt="Watchlist"/></span>
+                              <span onClick={()=>removeFromWatchlist(e.token_row_id)} ><img src="/assets/img/color.svg" alt="Watchlist"/></span>
                               :
-                              <span onClick={()=>addToWatchlist(e.token_row_id)} ><img src="/assets/img2/Star.svg" alt="Watchlist"/></span>
+                              <span onClick={()=>addToWatchlist(e.token_row_id)} ><img src="/assets/img/star.svg" alt="Watchlist"/></span>
                               }
                             </>
                             :
-                            <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url+"launchpad/upcoming"}><a onClick={()=> Logout()}><img src="/assets/img2/Star.svg" alt="Watchlist"/></a></Link>
+                            <Link href={app_coinpedia_url+"login?prev_url="+market_coinpedia_url+"launchpad/upcoming"}><a onClick={()=> Logout()}><img src="/assets/img/star.svg" alt="Watchlist"/></a></Link>
                           }
                           </td>
                           <td className="mobile_hide">{sl_no+i+1}</td>
                           <td className="mobile_td_fixed">
                             <a href={"/"+e.token_id}>
                               <div class="media">
+                              <div className='media-left align-self-center'>
                                 <img src={image_base_url+(e.token_image ? e.token_image : "default.png")} onError={(e) =>e.target.src = "/assets/img/default_token.png"} alt="Logo" />
-                                <div class="media-body">
+                                </div>
+                                <div class="media-body align-self-center">
                                   <h5 className="launchpad_token_title">{e.token_name} <span>{e.symbol}</span></h5>
                                 </div>
                               </div>
